Refactor: rest controller에서 restService로 openAPI 호출 코드 옮김
